+++ src/settings_docgen/main.py
@@ -1,10 +1,11 @@
 import importlib
+import re
 from collections.abc import Iterable as IterableCollection
 from enum import Enum, auto
 from functools import lru_cache
 from inspect import isclass
 from pathlib import Path
-from typing import Any, Type
+from typing import Any, List, Optional, Tuple, Type
 
 from click import Abort, BadParameter, secho
 from jinja2 import Environment, FileSystemLoader, select_autoescape
@@ -70,13 +71,43 @@ def main(
     ),
     output_format: OutputFormat = Option(..., "--output-format", "-f", help="Output format."),
     heading_offset: int = Option(0, min=0, help="How nested should be the top level heading generated."),
+    update_file: Optional[Path] = Option(
+        None,
+        "--update",
+        "-u",
+        exists=True,
+        writable=True,
+        file_okay=True,
+        dir_okay=False,
+        resolve_path=True,
+        help="Overwrite given file instead of writing to STDOUT. An error is raised if the "
+        "file doesn't exist or is not writable. Combine this flag with the '--between' "
+        "flag to update only a section of a file. ",
+    ),
+    update_between: Tuple[str, str] = Option(
+        (None, None),
+        "--between",
+        help="Update file given by '--update' between these two strings. New line after "
+        "the start mark/before the end mark is considered part of the pattern (if present). "
+        "Without the '--update' flag, this has no effect. ",
+    ),
+    templates: List[Path] = Option(
+        [],
+        exists=True,
+        readable=True,
+        file_okay=False,
+        dir_okay=True,
+        resolve_path=True,
+        help="One or more folders in a priority order to use when looking up templates for "
+        "generating output. Built-in templates will be used last if no matches found.",
+    ),
 ):
-    """Formats `pydantic.BaseSettings` into various formats."""
+    """Formats `pydantic.BaseSettings` into various formats. By default, the output is to STDOUT."""
     settings = import_class_path(class_path)
     render_kwargs = {"heading_offset": heading_offset, "fields": settings.__fields__.values()}
 
     env = Environment(
-        loader=FileSystemLoader(Path(__file__).parent / "templates"),
+        loader=FileSystemLoader(templates + [Path(__file__).parent / "templates"]),
         autoescape=select_autoescape(),
         trim_blocks=True,
         lstrip_blocks=True,
@@ -84,8 +115,30 @@ def main(
     )
     env.globals["is_values_with_descriptions"] = is_values_with_descriptions
     template = env.get_template(f"{output_format.value}.jinja")
+    render = template.render(**render_kwargs)
 
-    print(template.render(**render_kwargs))
+    if update_file is None:
+        print(render)
+        return
+
+    with open(update_file, "r", encoding="utf-8") as file:
+        content = file.read()
+
+    with open(update_file, "w", encoding="utf-8") as file:
+        if all(update_between):
+            pattern = f"({re.escape(update_between[0])}\n?).*(\n?{re.escape(update_between[1])})"
+            new_content = re.sub(pattern, f"\\1{render}\\2", content, count=1, flags=re.DOTALL)
+            if new_content == content:
+                secho(
+                    f"Boundary marks '{update_between[0]}' and '{update_between[1]}' not found in '{update_file}'. "
+                    f"Cannot update the content.",
+                    fg=colors.RED,
+                )
+                raise Abort()
+        else:
+            new_content = render
+
+        file.write(new_content)
 
 
 if __name__ == "__main__":